refactor: remove time-picker demo module
