+++ vaadin-time-picker-flow-parent/pom.xml
@@ -12,7 +12,6 @@
     <modules>
         <module>vaadin-time-picker-flow</module>
         <module>vaadin-time-picker-testbench</module>
-        <module>vaadin-time-picker-flow-demo</module>
     </modules>
     <profiles>
         <profile>
