--- vaadin-time-picker-flow-parent/vaadin-time-picker-flow-demo/src/main/resources/META-INF/resources/frontend/hide-clear-button-theme.html
@@ -1,9 +1,0 @@
-<dom-module theme-for="vaadin-combo-box" id="no-clear-button">
-    <template>
-        <style>
-            :host(.no-clear-button) [part="clear-button"]{
-                display:none
-            }
-        </style>
-    </template>
-</dom-module>
