+++ vaadin-time-picker-flow-parent/vaadin-time-picker-flow-integration-tests/src/test/java/com/vaadin/flow/component/timepicker/tests/TimePickerLocalizationIT.java
@@ -30,8 +30,8 @@ import org.junit.Test;
 import org.openqa.selenium.By;
 
 import com.vaadin.flow.component.timepicker.TimePicker;
-import com.vaadin.flow.component.timepicker.demo.TimePickerView;
 import com.vaadin.flow.component.timepicker.testbench.TimePickerElement;
+import com.vaadin.flow.component.timepicker.tests.TimePickerLocalizationView.LocalTimeTextBlock;
 import com.vaadin.tests.AbstractComponentIT;
 import com.vaadin.flow.testutil.TestPath;
 
@@ -390,7 +390,7 @@ public class TimePickerLocalizationIT extends AbstractComponentIT {
                 .replace("." + millisecondsInputValue, "");
 
         String[] splitLabelValue = getLabelValue()
-                .split(TimePickerView.LocalTimeTextBlock.MILLISECONDS_SPLIT);
+                .split(LocalTimeTextBlock.MILLISECONDS_SPLIT);
         String formattedTextValue = splitLabelValue[0].trim();
         String formattedTextValueMilliseconds = splitLabelValue[1].trim();
